start mocp -S at login, xfce4-appfinder shortcut
The mocp-scrobbler-script complains about moc not running so we just start it at login.

On some occasions I need something like a applications-menu (sometimes I forget how that
application is exactly called, especially when it is one I'm not using very often). So I
found xfce4-appfinder which acts as a menu.